server/routes/initialAppointment/handlers/hardStop/initialMeetingContact.ts
import { Request, Response } from 'express'
import LicenceService from '../../../../services/licenceService'
import UserType from '../../../../enumeration/userType'
import flashInitialApptUpdatedMessage from '../initialMeetingUpdatedFlashMessage'

export default class InitialMeetingContactRoutes {
  constructor(
    private readonly licenceService: LicenceService,
    private readonly userType: UserType
  ) {}

  GET = async (req: Request, res: Response): Promise<void> => {
    return res.render('pages/create/hardStop/initialMeetingContact')
  }

  POST = async (req: Request, res: Response): Promise<void> => {
    const { licenceId } = req.params
    const { user, licence } = res.locals
    await this.licenceService.updateContactNumber(licenceId, req.body, user)

    flashInitialApptUpdatedMessage(req, licence, this.userType)
    if (req.originalUrl?.includes('edit')) {
      res.redirect(`/licence/hard-stop/id/${licenceId}/check-your-answers`)
    } else {
      res.redirect(`/licence/hard-stop/create/id/${licenceId}/initial-meeting-time`)
    }
  }
}
